+++ src/main/java/vsla/userManager/company/CompanyServiceImpl.java
@@ -23,7 +23,7 @@ public class CompanyServiceImpl implements CompanyService {
     }
 
     @Override
-    public Company getCompanyById(Short companyId) {
+    public Company getCompanyById(Long companyId) {
         return companyRepository.findById(companyId)
                 .orElseThrow(() -> new ResourceNotFoundException("User not found"));
     }
